\name{primary_package-package}
\alias{primary_package-package}
\alias{primary_package}
\docType{package}
\title{
\packageTitle{primary_package}
}
\description{
\packageDescription{primary_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{primary_package}
\packageIndices{primary_package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{primary_package}

Maintainer: \packageMaintainer{primary_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
